Swtich to ascii theme for powerline with tmux
So I'm getting this weird issue where the powerline status line repeats
itself infinitely up the terminal.
This appears to be the result of either tmux or my terminal emulator
incorrectly guessing the width of the characters in question.

After much googlin' the best description of the problem I've found is a
bug report on powerline:
https://github.com/powerline/powerline/issues/1652

Unfortunately the powerline author, rightly or wrongly, has said it's
nothing to do with powerline.

I've yet to find a bug in gnome-terminal or tmux that covers the same
issue.

Switching to the ascii based theme for tmux should work for now