git-easy-worktree: Add initial commit
This was meant to be a nice simple script to manage worktrees slightly
easier.

The idea
========

The idea was the interface (after configuration) would be like this;

```console
$ git-easy-worktree create
```

and it would create the new worktree without asking any annoying
questions and also potentially run configured scripts after worktree is
created (like setting up venv).

Equally, removing the worktree would be;

```conosle
$ git-easy-worktree remove
```

and it would be removed, perhaps prompting if there are unsaved changes,
and again running pre-removal script.

Config would be repo specific and inherit in a similar way to the git
config.

Fucking CLI scripts
===================

I swear I spend my life writing the same shit over and over.

A CLI should be the easiest thing in the world, but I always hit the
same problems;

Trying to keep dependencies down. Keeping it simple.
----------------------------------------------------

If I'm just hacking a "quick" script I don't want to have to install a
bunch of other packages to make it work. Especially as these sometimes
get shared with colleagues and they just want to copy a file.

I know a variety of ways to bundle everything into a single file, but
would mean setting up CI and other stuff I cba with for a simple script.

Don't like any of the "standard" tools
--------------------------------------

Partly because of the above restriction on dependencies I feel stuck in
a loop writing the same config parsing, arg parsing, stuff over and over
again.

I've never really liked click for CLIs but it does come with some useful
tools for UI (prompts, styling etc).
I prefer arghparse which allows for the conveniant decorators for when
the script is simple, but doesn't tie my hands too much if it gets
complex which I always feel click does.
Also neither of these handle config files in a way I'd like.

Keep making the same mistakes
-----------------------------

Finally, because I don't have a consistent codebase that I can iterate
on, I make the same damn mistakes over and over, and when I solve a
problem in one place, it doesn't help me elsewhere.

Also GitPython
==============

Is still a pain. It's improved a lot, but it's pretty painful for doing
anything with the config which meant I ended up having to call out to
git with subprocess.

Also, I underestimated the limitation of gits ini-style config for
multiline strings. This meant I have to combine a config in yaml with
git config which has made a right clusterfuck.

General Hackiness
=================

Because this whole thing took longer than I wanted, it's both incomplete
and hacky as all hell. Both of which are irritating. Balls.